fix routeable modal registration so it becomes a proper controller
