+++ src/packages/core/modal/index.ts
@@ -1,7 +1,7 @@
 import './component/modal.element.js';
 
 export * from './context/index.js';
-export * from './route-registration/modal-route-registration.js';
+export * from './route-registration/modal-route-registration.interface.js';
 export * from './route-registration/modal-route-registration.controller.js';
 export * from './token/index.js';
 export * from './types.js';
